api/requirements.txt
fastapi
google-genai
pydantic
uvicorn
httpx
python-multipart
